Fixes verification of data signed by a signature using a DSA algorithm
The -PGPSignaturePacket verifyData:publicKey:... method was validating the data to be hashed, rather than the hashed data.

The PGPDSA verify:... method was failing when checking missing MPI values.   specifically failing when the MPI R and S values were set  rather that not set.

This fix also adds a verify:....error: method to allow the pass back of issues with DSA verification, and adds a PGPErrorSignatureVerificationFailure value to the PGPErrorCode enumeration